third_party/blink/web_tests/svg/filters/feConvolveMatrix-error-handling-expected.txt

CONSOLE ERROR: Error: <feConvolveMatrix> attribute order: A negative value is not valid. ("-1 -1")
